requirements.txt
decord==0.6.0
einops==0.3.0
imageio==2.9.0
numpy==1.24.2
omegaconf==2.1.1
opencv_python
pandas==2.0.0
Pillow==9.5.0
PyYAML==6.0
pydantic==1.10.13
setuptools==65.6.3
torch==2.2.0
torchvision
pytorch_lightning==1.9.3
tqdm==4.65.0
transformers==4.36.0
fairscale==0.4.13
moviepy
av
xformers==0.0.24
#https://nebula-cv.oss-cn-zhangjiakou.aliyuncs.com/docker_image/ppu/sdk_1_3/torch_230/xformers-0.0.22-cp310-cp310-linux_x86_64.whl
timm==0.4.12
scikit-learn 
open_clip_torch==2.22.0
kornia
matplotlib
tensorboard
scipy
opencv-python
yacs
loguru
